refactor: renames, lint
toVdoc -> vdon
